Moved controllers related to new submissions
The are now in submit/controllers/ui/new
Tests related to controllers also fixed.

ARXIVNG-2898
