+++ submit/controllers/ui/new/license.py
@@ -19,8 +19,8 @@ from arxiv.license import LICENSES
 from arxiv.users.domain import Session
 from arxiv.submission import save, InvalidEvent, SaveError
 from arxiv.submission.domain.event import SetLicense
-from ...util import load_submission
-from .util import validate_command, user_and_client_from_session
+from submit.util import load_submission
+from submit.controllers.ui.util import validate_command, user_and_client_from_session
 
 
 logger = logging.getLogger(__name__)  # pylint: disable=C0103
