+++ submit/controllers/ui/new/process.py
@@ -33,8 +33,8 @@ from arxiv.submission.domain.preview import Preview
 from arxiv.submission.domain.submission import Compilation, SubmissionContent
 from arxiv.submission.services import PreviewService, Compiler
 
-from ...util import load_submission
-from .util import validate_command, user_and_client_from_session
+from submit.util import load_submission
+from submit.controllers.ui.util import validate_command, user_and_client_from_session
 
 logger = logging.getLogger(__name__)
