+++ submit/controllers/ui/new/tests/test_authorship.py
@@ -1,4 +1,4 @@
-"""Tests for :mod:`submit.controllers.policy`."""
+"""Tests for :mod:`submit.controllers.authorship`."""
 
 from datetime import timedelta, datetime
 from http import HTTPStatus as status
@@ -10,13 +10,14 @@ from werkzeug.exceptions import InternalServerError, NotFound, BadRequest
 from wtforms import Form
 
 import arxiv.submission as events
-from arxiv.submission.domain.event import ConfirmPolicy
+from arxiv.submission.domain.event import ConfirmAuthorship
 from arxiv.users import auth, domain
-from submit.controllers import policy
+
+from submit.controllers.ui.new import authorship
 
 
-class TestConfirmPolicy(TestCase):
-    """Test behavior of :func:`.policy` controller."""
+class TestVerifyAuthorship(TestCase):
+    """Test behavior of :func:`.authorship` controller."""
 
     def setUp(self):
         """Create an authenticated session."""
@@ -48,22 +49,20 @@ class TestConfirmPolicy(TestCase):
             )
         )
 
-    @mock.patch(f'{policy.__name__}.PolicyForm.Meta.csrf', False)
+    @mock.patch(f'{authorship.__name__}.AuthorshipForm.Meta.csrf', False)
     @mock.patch('arxiv.submission.load')
     def test_get_request_with_submission(self, mock_load):
         """GET request with a submission ID."""
         submission_id = 2
         before = mock.MagicMock(submission_id=submission_id,
-                                is_finalized=False,
-                                submitter_accepts_policy=False)
+                                submitter_is_author=False)
         mock_load.return_value = (before, [])
-        data = MultiDict()
-
-        data, code, _ = policy.policy('GET', data, self.session, submission_id)
+        data, code, _ = authorship.authorship('GET', MultiDict(), self.session,
+                                              submission_id)
         self.assertEqual(code, status.OK, "Returns 200 OK")
         self.assertIsInstance(data['form'], Form, "Data includes a form")
 
-    @mock.patch(f'{policy.__name__}.PolicyForm.Meta.csrf', False)
+    @mock.patch(f'{authorship.__name__}.AuthorshipForm.Meta.csrf', False)
     @mock.patch('arxiv.submission.load')
     def test_get_request_with_nonexistant_submission(self, mock_load):
         """GET request with a submission ID."""
@@ -73,88 +72,81 @@ class TestConfirmPolicy(TestCase):
             raise events.exceptions.NoSuchSubmission('Nada')
 
         mock_load.side_effect = raise_no_such_submission
-        with self.assertRaises(NotFound):
-            policy.policy('GET', MultiDict(), self.session,  submission_id)
+        params = MultiDict()
 
-    @mock.patch(f'{policy.__name__}.PolicyForm.Meta.csrf', False)
+        with self.assertRaises(NotFound):
+            authorship.authorship('GET', params, self.session, submission_id)
+
+    @mock.patch(f'{authorship.__name__}.AuthorshipForm.Meta.csrf', False)
     @mock.patch('arxiv.submission.load')
     def test_post_request(self, mock_load):
         """POST request with no data."""
         submission_id = 2
         before = mock.MagicMock(submission_id=submission_id,
-                                is_finalized=False,
-                                submitter_accepts_policy=False)
+                                submitter_is_author=False)
         mock_load.return_value = (before, [])
-
         params = MultiDict()
-        try:
-            policy.policy('POST', params, self.session, submission_id)
-            self.fail('BadRequest was not raised')
-        except BadRequest as e:
-            data = e.description
-            self.assertIsInstance(data['form'], Form, "Data includes a form")
+        with self.assertRaises(BadRequest):
+            authorship.authorship('POST', params, self.session, submission_id)
 
-    @mock.patch(f'{policy.__name__}.PolicyForm.Meta.csrf', False)
+    @mock.patch(f'{authorship.__name__}.AuthorshipForm.Meta.csrf', False)
     @mock.patch('arxiv.submission.load')
     def test_not_author_no_proxy(self, mock_load):
         """User indicates they are not author, but also not proxy."""
         submission_id = 2
         before = mock.MagicMock(submission_id=submission_id,
-                                is_finalized=False,
-                                submitter_accepts_policy=False)
+                                submitter_is_author=False)
         mock_load.return_value = (before, [])
-        params = MultiDict({})
-        try:
-            policy.policy('POST', params, self.session, submission_id)
-            self.fail('BadRequest was not raised')
-        except BadRequest as e:
-            data = e.description
-            self.assertIsInstance(data['form'], Form, "Data includes a form")
+        params = MultiDict({'authorship': authorship.AuthorshipForm.NO})
 
-    @mock.patch(f'{policy.__name__}.PolicyForm.Meta.csrf', False)
-    @mock.patch('submit.controllers.util.url_for')
-    @mock.patch(f'{policy.__name__}.save')
+        with self.assertRaises(BadRequest):
+            authorship.authorship('POST', params, self.session, submission_id)
+
+    @mock.patch(f'{authorship.__name__}.AuthorshipForm.Meta.csrf', False)
+    @mock.patch('submit.controllers.ui.util.url_for')
+    @mock.patch(f'{authorship.__name__}.save')
     @mock.patch('arxiv.submission.load')
     def test_post_request_with_data(self, mock_load, mock_save, mock_url_for):
-        """POST request with `policy` set."""
+        """POST request with `authorship` set."""
         # Event store does not complain; returns object with `submission_id`.
         submission_id = 2
         before = mock.MagicMock(submission_id=submission_id,
                                 is_finalized=False,
-                                submitter_accepts_policy=False)
+                                submitter_is_author=False)
         after = mock.MagicMock(submission_id=submission_id, is_finalized=False)
         mock_load.return_value = (before, [])
         mock_save.return_value = (after, [])
         mock_url_for.return_value = 'https://foo.bar.com/yes'
 
-        params = MultiDict({'policy': 'y', 'action': 'next'})
-        _, code, _ = policy.policy('POST', params, self.session, submission_id)
+        params = MultiDict({'authorship': 'y', 'action': 'next'})
+        _, code, _ = authorship.authorship('POST', params, self.session,
+                                           submission_id)
         self.assertEqual(code, status.SEE_OTHER, "Returns redirect")
 
-    @mock.patch(f'{policy.__name__}.PolicyForm.Meta.csrf', False)
-    @mock.patch('submit.controllers.util.url_for')
-    @mock.patch(f'{policy.__name__}.save')
+    @mock.patch(f'{authorship.__name__}.AuthorshipForm.Meta.csrf', False)
+    @mock.patch('submit.controllers.ui.util.url_for')
+    @mock.patch(f'{authorship.__name__}.save')
     @mock.patch('arxiv.submission.load')
     def test_save_fails(self, mock_load, mock_save, mock_url_for):
-        """Event store flakes out on saving policy acceptance."""
+        """Event store flakes out on saving the command."""
         submission_id = 2
         before = mock.MagicMock(submission_id=submission_id,
                                 is_finalized=False,
-                                submitter_accepts_policy=False)
+                                submitter_is_author=False)
         mock_load.return_value = (before, [])
 
-        # Event store does not complain; returns object with `submission_id`
-        def raise_on_policy(*ev, **kwargs):
-            if type(ev[0]) is ConfirmPolicy:
-                raise events.SaveError('the end of the world as we know it')
-            ident = kwargs.get('submission_id', 2)
-            return (mock.MagicMock(submission_id=ident), [])
+        def raise_on_verify(*ev, **kwargs):
+            if type(ev[0]) is ConfirmAuthorship:
+                raise events.SaveError('The world is ending')
+            submission_id = kwargs.get('submission_id', 2)
+            return (mock.MagicMock(submission_id=submission_id), [])
 
-        mock_save.side_effect = raise_on_policy
-        params = MultiDict({'policy': 'y', 'action': 'next'})
+        mock_save.side_effect = raise_on_verify
+        params = MultiDict({'authorship': 'y', 'action': 'next'})
+
         try:
-            policy.policy('POST', params, self.session, 2)
+            authorship.authorship('POST', params, self.session, 2)
             self.fail('InternalServerError not raised')
         except InternalServerError as e:
             data = e.description
-            self.assertIsInstance(data['form'], Form, "Data includes a form")
+            self.assertIsInstance(data['form'], Form, "Data includes form")
