+++ submit/controllers/ui/new/tests/test_policy.py
@@ -1,4 +1,4 @@
-"""Tests for :mod:`submit.controllers.authorship`."""
+"""Tests for :mod:`submit.controllers.policy`."""
 
 from datetime import timedelta, datetime
 from http import HTTPStatus as status
@@ -10,14 +10,13 @@ from werkzeug.exceptions import InternalServerError, NotFound, BadRequest
 from wtforms import Form
 
 import arxiv.submission as events
-from arxiv.submission.domain.event import ConfirmAuthorship
+from arxiv.submission.domain.event import ConfirmPolicy
 from arxiv.users import auth, domain
+from submit.controllers.ui.new import policy
 
-from submit.controllers import authorship
 
-
-class TestVerifyAuthorship(TestCase):
-    """Test behavior of :func:`.authorship` controller."""
+class TestConfirmPolicy(TestCase):
+    """Test behavior of :func:`.policy` controller."""
 
     def setUp(self):
         """Create an authenticated session."""
@@ -49,20 +48,22 @@ class TestVerifyAuthorship(TestCase):
             )
         )
 
-    @mock.patch(f'{authorship.__name__}.AuthorshipForm.Meta.csrf', False)
+    @mock.patch(f'{policy.__name__}.PolicyForm.Meta.csrf', False)
     @mock.patch('arxiv.submission.load')
     def test_get_request_with_submission(self, mock_load):
         """GET request with a submission ID."""
         submission_id = 2
         before = mock.MagicMock(submission_id=submission_id,
-                                submitter_is_author=False)
+                                is_finalized=False,
+                                submitter_accepts_policy=False)
         mock_load.return_value = (before, [])
-        data, code, _ = authorship.authorship('GET', MultiDict(), self.session,
-                                              submission_id)
+        data = MultiDict()
+
+        data, code, _ = policy.policy('GET', data, self.session, submission_id)
         self.assertEqual(code, status.OK, "Returns 200 OK")
         self.assertIsInstance(data['form'], Form, "Data includes a form")
 
-    @mock.patch(f'{authorship.__name__}.AuthorshipForm.Meta.csrf', False)
+    @mock.patch(f'{policy.__name__}.PolicyForm.Meta.csrf', False)
     @mock.patch('arxiv.submission.load')
     def test_get_request_with_nonexistant_submission(self, mock_load):
         """GET request with a submission ID."""
@@ -72,80 +73,88 @@ class TestVerifyAuthorship(TestCase):
             raise events.exceptions.NoSuchSubmission('Nada')
 
         mock_load.side_effect = raise_no_such_submission
-        params = MultiDict()
-
         with self.assertRaises(NotFound):
-            authorship.authorship('GET', params, self.session, submission_id)
+            policy.policy('GET', MultiDict(), self.session,  submission_id)
 
-    @mock.patch(f'{authorship.__name__}.AuthorshipForm.Meta.csrf', False)
+    @mock.patch(f'{policy.__name__}.PolicyForm.Meta.csrf', False)
     @mock.patch('arxiv.submission.load')
     def test_post_request(self, mock_load):
         """POST request with no data."""
         submission_id = 2
         before = mock.MagicMock(submission_id=submission_id,
-                                submitter_is_author=False)
+                                is_finalized=False,
+                                submitter_accepts_policy=False)
         mock_load.return_value = (before, [])
-        params = MultiDict()
-        with self.assertRaises(BadRequest):
-            authorship.authorship('POST', params, self.session, submission_id)
 
-    @mock.patch(f'{authorship.__name__}.AuthorshipForm.Meta.csrf', False)
+        params = MultiDict()
+        try:
+            policy.policy('POST', params, self.session, submission_id)
+            self.fail('BadRequest was not raised')
+        except BadRequest as e:
+            data = e.description
+            self.assertIsInstance(data['form'], Form, "Data includes a form")
+
+    @mock.patch(f'{policy.__name__}.PolicyForm.Meta.csrf', False)
     @mock.patch('arxiv.submission.load')
     def test_not_author_no_proxy(self, mock_load):
         """User indicates they are not author, but also not proxy."""
         submission_id = 2
         before = mock.MagicMock(submission_id=submission_id,
-                                submitter_is_author=False)
+                                is_finalized=False,
+                                submitter_accepts_policy=False)
         mock_load.return_value = (before, [])
-        params = MultiDict({'authorship': authorship.AuthorshipForm.NO})
+        params = MultiDict({})
+        try:
+            policy.policy('POST', params, self.session, submission_id)
+            self.fail('BadRequest was not raised')
+        except BadRequest as e:
+            data = e.description
+            self.assertIsInstance(data['form'], Form, "Data includes a form")
 
-        with self.assertRaises(BadRequest):
-            authorship.authorship('POST', params, self.session, submission_id)
-
-    @mock.patch(f'{authorship.__name__}.AuthorshipForm.Meta.csrf', False)
-    @mock.patch('submit.controllers.util.url_for')
-    @mock.patch(f'{authorship.__name__}.save')
+    @mock.patch(f'{policy.__name__}.PolicyForm.Meta.csrf', False)
+    @mock.patch('submit.controllers.ui.util.url_for')
+    @mock.patch(f'{policy.__name__}.save')
     @mock.patch('arxiv.submission.load')
     def test_post_request_with_data(self, mock_load, mock_save, mock_url_for):
-        """POST request with `authorship` set."""
+        """POST request with `policy` set."""
         # Event store does not complain; returns object with `submission_id`.
         submission_id = 2
         before = mock.MagicMock(submission_id=submission_id,
                                 is_finalized=False,
-                                submitter_is_author=False)
+                                submitter_accepts_policy=False)
         after = mock.MagicMock(submission_id=submission_id, is_finalized=False)
         mock_load.return_value = (before, [])
         mock_save.return_value = (after, [])
         mock_url_for.return_value = 'https://foo.bar.com/yes'
 
-        params = MultiDict({'authorship': 'y', 'action': 'next'})
-        _, code, _ = authorship.authorship('POST', params, self.session,
-                                           submission_id)
+        params = MultiDict({'policy': 'y', 'action': 'next'})
+        _, code, _ = policy.policy('POST', params, self.session, submission_id)
         self.assertEqual(code, status.SEE_OTHER, "Returns redirect")
 
-    @mock.patch(f'{authorship.__name__}.AuthorshipForm.Meta.csrf', False)
-    @mock.patch('submit.controllers.util.url_for')
-    @mock.patch(f'{authorship.__name__}.save')
+    @mock.patch(f'{policy.__name__}.PolicyForm.Meta.csrf', False)
+    @mock.patch('submit.controllers.ui.util.url_for')
+    @mock.patch(f'{policy.__name__}.save')
     @mock.patch('arxiv.submission.load')
     def test_save_fails(self, mock_load, mock_save, mock_url_for):
-        """Event store flakes out on saving the command."""
+        """Event store flakes out on saving policy acceptance."""
         submission_id = 2
-        before = mock.MagicMock(submission_id=submission_id, is_finalized=False,
-                                submitter_is_author=False)
+        before = mock.MagicMock(submission_id=submission_id,
+                                is_finalized=False,
+                                submitter_accepts_policy=False)
         mock_load.return_value = (before, [])
 
-        def raise_on_verify(*ev, **kwargs):
-            if type(ev[0]) is ConfirmAuthorship:
-                raise events.SaveError('The world is ending')
-            submission_id = kwargs.get('submission_id', 2)
-            return (mock.MagicMock(submission_id=submission_id), [])
+        # Event store does not complain; returns object with `submission_id`
+        def raise_on_policy(*ev, **kwargs):
+            if type(ev[0]) is ConfirmPolicy:
+                raise events.SaveError('the end of the world as we know it')
+            ident = kwargs.get('submission_id', 2)
+            return (mock.MagicMock(submission_id=ident), [])
 
-        mock_save.side_effect = raise_on_verify
-        params = MultiDict({'authorship': 'y', 'action': 'next'})
-
+        mock_save.side_effect = raise_on_policy
+        params = MultiDict({'policy': 'y', 'action': 'next'})
         try:
-            authorship.authorship('POST', params, self.session, 2)
+            policy.policy('POST', params, self.session, 2)
             self.fail('InternalServerError not raised')
         except InternalServerError as e:
             data = e.description
-            self.assertIsInstance(data['form'], Form, "Data includes form")
+            self.assertIsInstance(data['form'], Form, "Data includes a form")
