+++ submit/controllers/ui/new/verify_user.py
@@ -18,8 +18,8 @@ from arxiv.users.domain import Session
 from arxiv.submission import save, SaveError
 from arxiv.submission.domain.event import ConfirmContactInformation
 
-from ...util import load_submission
-from .util import validate_command, user_and_client_from_session
+from submit.util import load_submission
+from submit.controllers.ui.util import validate_command, user_and_client_from_session
 
 
 logger = logging.getLogger(__name__)    # pylint: disable=C0103
