+++ submit/filters/__init__.py
@@ -10,7 +10,7 @@ from arxiv.submission.domain.process import ProcessStatus
 from arxiv.submission.domain.submission import Compilation
 from arxiv.submission.domain.uploads import FileStatus
 
-from submit.controllers.ui.upload import group_files
+from submit.controllers.ui.new.upload import group_files
 from submit.util import tidy_filesize
 
 from .tex_filters import compilation_log_display
